wip: starts conversion to 4-series
