--- .github/workflows/essentialsplugins-releasebuilds-caller.yml
@@ -1,26 +1,0 @@
-
-name: Release Build
-
-on:
-  release:
-    types:
-      - released
-    branches:
-      - "main"
-
-  workflow_dispatch: 
-    inputs: 
-      branch:
-        description: 'Branch to build'
-        required: true 
-        type: string
-
-jobs:
-  call-workflow:
-    uses: PepperDash/workflow-templates/.github/workflows/essentialsplugins-builds.yml@main
-    secrets: inherit
-    with: 
-      branch: "main"
-      default-branch: "main"
-
-
