+++ epi-panasonic-camera/epi-panasonic-camera.4series.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.2.32630.192
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "epi-panasonic-camera", "epi-panasonic-camera\epi-panasonic-camera.csproj", "{2DED7778-602C-4A9B-B2ED-8DF4FEF65FF7}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2DED7778-602C-4A9B-B2ED-8DF4FEF65FF7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2DED7778-602C-4A9B-B2ED-8DF4FEF65FF7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2DED7778-602C-4A9B-B2ED-8DF4FEF65FF7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2DED7778-602C-4A9B-B2ED-8DF4FEF65FF7}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {669966C2-74FC-4EF1-9A37-EE2746879D65}
+	EndGlobalSection
+EndGlobal
