+++ epi-panasonic-camera/epi-panasonic-camera/epi-panasonic-camera.4series.csproj
@@ -1,0 +1,8 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net472</TargetFramework>
+    <RootNamespace>PepperDash.Essentials.Plugins.Camera.Panasonic</RootNamespace>
+  </PropertyGroup>
+
+</Project>
