+++ src/Directory.Build.props
@@ -1,0 +1,21 @@
+﻿<Project>
+  <PropertyGroup>
+    <Version>1.0.0-local</Version>
+    <Authors>PepperDash Technologies</Authors>
+    <Company>PepperDash Technologies</Company>
+    <Product>PepperDash Essentials Panasonic Camera</Product>    
+    <RepositoryUrl>https://github.com/PepperDash/epi-panasonic-camera-http</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+    <PackageTags>Crestron; 4series</PackageTags>
+    <PackageOutputPath>../output</PackageOutputPath>
+    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+    <PackageLicenseFile>LICENSE.md</PackageLicenseFile>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
+  </PropertyGroup>
+  <ItemGroup>
+    <None Include="..\LICENSE.md" Pack="true" PackagePath=""/>
+    <None Include="..\README.md" Pack="true" PackagePath=""/>
+    <None Include=".\deviceTypeMetadata.json" Pack="true" PackagePath="content;"/>
+    <None Include=".\deviceTypePropertiesSchema.json" Pack="true" PackagePath="content;"/>
+  </ItemGroup>
+</Project>
