+++ src/PanasonicCameraExtensions.cs
