+++ src/PanasonicCameraMonitor.cs
