+++ src/PanasonicCameraPropsConfig.cs
