+++ src/Properties/AssemblyInfo.cs
