+++ src/epi-panasonic-camera.nuspec
